server/requirements.txt
fastapi==0.115.5
uvicorn==0.30.6
pandas==2.2.3
numpy==2.1.3
openpyxl==3.1.5
openai==1.57.1
python-dateutil==2.9.0.post0
python-dotenv>=1.0.0
# LangChain dependencies for RAG and web search
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-dashscope>=0.0.1
langchain-core>=0.1.0
langgraph>=0.0.20
faiss-cpu>=1.7.4
tavily-python>=0.3.0
# Document loaders dependencies
pypdf>=3.0.0
docx2txt>=0.8
# Chart generation dependencies
matplotlib>=3.7.0
# PDF generation dependencies (optional)
playwright>=1.40.0

